requirements.txt
flask==3.0.2
gunicorn==21.2.0
python-dotenv==1.0.1
requests==2.31.0